@@ -3,9 +3,9 @@ import socket
 from socket import SOCK_DGRAM
 from unittest.mock import Mock, create_autospec, patch
 
-from greeclimate.network import DeviceProtocol2
+from greeclimate.network import DeviceProtocolBase2
 
-DEFAULT_TIMEOUT = 5
+DEFAULT_TIMEOUT = 1
 DISCOVERY_REQUEST = {"t": "scan"}
 DISCOVERY_RESPONSE = {
     "t": "pack",
@@ -51,14 +51,33 @@ DISCOVERY_RESPONSE_NO_CID = {
         "lock": 0,
     },
 }
+DEFAULT_REQUEST = {
+    "t": "pack",
+    "i": 1,
+    "uid": 0,
+    "cid": "aabbcc112233",
+    "tcid": "",
+    "pack": {
+        "t": "test"
+    }
+}
 DEFAULT_RESPONSE = {
     "t": "pack",
     "i": 1,
     "uid": 0,
     "cid": "aabbcc112233",
     "tcid": "",
-    "pack": {},
+    "pack": {
+        "t": "testresponse"
+    }
 }
+
+
+def generate_response(data):
+    """Generate a response from a request."""
+    response = DEFAULT_RESPONSE.copy()
+    response["pack"].update(data)
+    return response
 
 
 def get_mock_device_info():
@@ -76,24 +95,25 @@ def get_mock_device_info():
 def encrypt_payload(data):
     """Encrypt the payload of responses quickly."""
     d = data.copy()
-    d["pack"] = DeviceProtocol2.encrypt_payload(d["pack"])
+    d["pack"] = DeviceProtocolBase2.encrypt_payload(d["pack"])
     return d
 
 
 class Responder:
     """Context manage for easy raw socket responders."""
 
-    def __init__(self, family, addr) -> None:
+    def __init__(self, family, addr, bcast=False) -> None:
         """Initialize the class."""
         self.sock = None
         self.family = family
         self.addr = addr
+        self.bcast = bcast
 
     def __enter__(self):
         """Enter the context manager."""
         self.sock = socket.socket(self.family, SOCK_DGRAM)
         self.sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        self.sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
+        self.sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, self.bcast)
         self.sock.settimeout(DEFAULT_TIMEOUT)
         self.sock.bind(("", self.addr))
         return self.sock
